requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
djangorestframework==3.14.0
pycodestyle==2.9.1
pytz==2022.4
sqlparse==0.4.3
toml==0.10.2
